backend/requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
sqlalchemy>=2.0.0
pydantic>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
aiofiles>=23.1.0
langchain>=0.0.267
openai>=1.0.0
pinecone-client>=2.2.1
pytesseract>=0.3.10
easyocr>=1.7.0
pdf2image>=1.16.3
python-dotenv>=1.0.0
pytest>=7.3.1
httpx>=0.24.0
tenacity>=8.2.2
pillow>=9.5.0
